Add task DTOs
