+++ server/__init__.py
@@ -75,16 +75,15 @@ def init_flask_restful_routes(app):
     api = Api(app, default_mediatype='application/json')
 
     from server.api.health_check_api import HealthCheckAPI
-    from server.api.mapping_apis import ProjectAPI, TaskAPI, LockTaskForMappingAPI
+    from server.api.mapping_apis import ProjectAPI, TaskAPI, LockTaskForMappingAPI, UnlockTaskForMappingAPI
     from server.api.project_admin_api import ProjectAdminAPI
     from server.api.swagger_docs_api import SwaggerDocsAPI
-    from server.api.tasks_api import UnlockTaskAPI
 
-    api.add_resource(SwaggerDocsAPI,        '/api/docs')
-    api.add_resource(HealthCheckAPI,        '/api/health-check')
-    api.add_resource(ProjectAPI,            '/api/v1/project/<int:project_id>')
-    api.add_resource(LockTaskForMappingAPI, '/api/v1/project/<int:project_id>/task/<int:task_id>/lock-for-mapping')
-    api.add_resource(ProjectAdminAPI,       '/api/v1/admin/project', endpoint="create_project", methods=['PUT'])
-    api.add_resource(ProjectAdminAPI,       '/api/v1/admin/project/<int:project_id>', methods=['GET', 'POST'])
-    api.add_resource(TaskAPI,               '/api/v1/project/<int:project_id>/task/<int:task_id>')
-    api.add_resource(UnlockTaskAPI,         '/api/v1/project/<int:project_id>/task/<int:task_id>/unlock')
+    api.add_resource(SwaggerDocsAPI,          '/api/docs')
+    api.add_resource(HealthCheckAPI,          '/api/health-check')
+    api.add_resource(ProjectAPI,              '/api/v1/project/<int:project_id>')
+    api.add_resource(LockTaskForMappingAPI,   '/api/v1/project/<int:project_id>/task/<int:task_id>/lock-for-mapping')
+    api.add_resource(ProjectAdminAPI,         '/api/v1/admin/project', endpoint="create_project", methods=['PUT'])
+    api.add_resource(ProjectAdminAPI,         '/api/v1/admin/project/<int:project_id>', methods=['GET', 'POST'])
+    api.add_resource(TaskAPI,                 '/api/v1/project/<int:project_id>/task/<int:task_id>')
+    api.add_resource(UnlockTaskForMappingAPI, '/api/v1/project/<int:project_id>/task/<int:task_id>/unlock-after-mapping')
